% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTimePoints.R
\name{as.data.frame.TP}
\alias{as.data.frame.TP}
\title{Function for converting an object of class TP to a data.frame.}
\usage{
\method{as.data.frame}{TP}(x, ...)
}
\arguments{
\item{x}{An object of class TP.}

\item{...}{Ignored.}
}
\description{
Function for converting an object of class TP to a data.frame.
}
