% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chkMissing}
\alias{chkMissing}
\title{Helper function for checking missing values for a given trait.}
\usage{
chkMissing(TP, trait)
}
\arguments{
\item{TP}{An object of class TP.}

\item{trait}{A character string indicating the trait for which missings
should be checked.}
}
\description{
Helper function for checking missing values for a given trait.
}
\keyword{internal}
