% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctSpatial.R
\name{correctSpatial}
\alias{correctSpatial}
\title{Approach 2: correct for spatial effects and other unnecesary factors}
\usage{
correctSpatial(fitMod)
}
\description{
Approach 2: correct for spatial effects and other unnecesary factors
}
\keyword{internal}
