% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{countValidPlot}
\alias{countValidPlot}
\title{Count valid observations per plotId for a given trait}
\usage{
countValidPlot(TP, trait, plotIds = NULL)
}
\arguments{
\item{TP}{An object of class TP.}

\item{trait}{A character string indicating the trait for which valid
observations should be counted.}

\item{plotIds}{A character vector indicating the plotIds for which valid
observations should be checked. If \code{NULL} valid observations are
counted for all plotIds in TP.}
}
\description{
Count valid observations per plotId for a given trait.
}
