% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTimePoints.R
\name{createTimePoints}
\alias{createTimePoints}
\title{Create an object of class TP}
\usage{
createTimePoints(dat, genotype, timePoint, plotId = NULL, repId = NULL,
  rowNum = NULL, colNum = NULL, rowId = rowNum, colId = colNum,
  addCheck = FALSE, checkGenotypes = NULL)
}
\arguments{
\item{dat}{A data.frame.}

\item{genotype}{A character string indicating the column in dat containing
the genotypes.}

\item{timePoint}{A character string indicating the column in dat containing
the time points.}

\item{plotId}{A character string indicating the column in dat containing
the plotId. This has to be a unique identifier per plot.}

\item{repId}{A character string indicating the column in dat containing
the replicates.}

\item{rowNum}{A character string indicating the column in dat containing
the row number of the plot.}

\item{colNum}{A character string indicating the column in dat containing
the column number of the plot.}

\item{rowId}{A character string indicating the column in dat containing
the row number of the plot.}

\item{colId}{A character string indicating the column in dat containing
the column number of the plot.}

\item{addCheck}{Should a column check be added to the output. If \code{TRUE}
checkGenotypes cannot be \code{NULL}.}

\item{checkGenotypes}{A character vector containing the genotypes used as
checks in the experiment.}
}
\value{
An object of class TP. A list with per time point in the input a
data.frame containing the data for that time point.
}
\description{
Create an object of class TP
}
