% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectOutliers.R
\name{detectOutliers}
\alias{detectOutliers}
\title{Outlier detection}
\usage{
detectOutliers(
  corrDat,
  predDat,
  coefDat,
  trait,
  genotypes = unique(corrDat[["genotype"]]),
  thrCor = 0.9,
  thrPca = 1,
  title = NULL,
  outFile = NULL,
  outFileOpts = NULL
)
}
\arguments{
\item{corrDat}{A data.frame with corrected spatial data.}

\item{predDat}{A data.frame with predicted data from a fitted spline.}

\item{coefDat}{A data.frame with coefficients from a fitted spline.}

\item{trait}{A character string indicating the trait for which to detect
outliers.}

\item{genotypes}{A character vector indicating the genotypes for which to
detect outliers.}

\item{thrCor}{A numerical value used as threshold for determining outliers
based on correlation between plots.}

\item{thrPca}{A numerical value used as threshold for determining outliers
based on PCA scores.}

\item{title}{A character string, the main title added to all plots in the
output.}

\item{outFile}{A character string indicating the .csv file to which the
results should be written. If \code{NULL} no file is written and all plots
are made within R. Warning: this is potentially very slow for large numbers
of genotypes.}

\item{outFileOpts}{A named list of extra options for the pdf output file,
e.g. width and height. See \code{\link[grDevices]{pdf}} for all possible
options.}
}
\description{
Function for detecting outliers.
}
