% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectSingleOut.R
\name{detectSingleOut}
\alias{detectSingleOut}
\title{Detect outliers in a time series}
\usage{
detectSingleOut(
  TP,
  trait,
  plotIds = NULL,
  checkEdges = TRUE,
  confIntSize = 5,
  nnLocfit = 0.5
)
}
\arguments{
\item{TP}{An object of class TP.}

\item{trait}{A character vector indicating the trait to model in TP.}

\item{plotIds}{A character vector of plotIds for which the outliers should be
detected. If \code{NULL}, all plotId in TP are used.}

\item{checkEdges}{Before fitting the local regression should a check be done
if the first and last time point for a plot are outlying observations?}

\item{confIntSize}{A numeric value defining the confidence interval.}

\item{nnLocfit}{A numeric value defining the constant component of the
smoothing parameter nn. (see the locfit())}
}
\value{
An object of class singleOut, a data.frame with the following
columns.
\describe{
\item{plotId}{the id variable}
\item{timePoint}{time variable in datain}
\item{trait}{name of the modeled variable in datain}
\item{yPred}{the locfit prediction}
\item{sd_yPred}{standard deviation of the prediction}
\item{lwr}{lower bound of the confidence interval}
\item{upr}{upper bound of the confidence interval}
\item{outlier}{flag of detected outlier (0 is outlier, 1 is not)}
}
}
\description{
Detect outlying observations in a time series by modeling each plotId using
a local regression.
}
\details{
See locfit() help function from the locfit R library. The user can act on:
\describe{
\item{nnLocfit}{the constant of the smoothing parameter. Increase nnLocfit
to have a very smooth curve}
\item{confIntSize}{the level to calculate the confidence interval. Increase
confIntSize to exclude less outliers}
}
}
\examples{
## Create a TP object containing the data from the Phenovator.
PhenovatorDat1 <- PhenovatorDat1[!PhenovatorDat1$pos \%in\%
                                 c("c24r41", "c7r18", "c7r49"), ]
phenoTP <- createTimePoints(dat = PhenovatorDat1,
                            experimentName = "Phenovator",
                            genotype = "Genotype",
                            timePoint = "timepoints",
                            repId = "Replicate",
                            plotId = "pos",
                            rowNum = "y", colNum = "x",
                            addCheck = TRUE,
                            checkGenotypes = c("check1", "check2",
                                               "check3", "check4"))

## First select a subset of plants, for example here 9 plants
plantSel <- phenoTP[[1]]$plotId[1:9]
# Then run on the subset
resuVatorHTP <- detectSingleOut (TP = phenoTP,
                                 trait = "EffpsII",
                                 plotIds = plantSel,
                                 confIntSize = 3,
                                 nnLocfit = 0.1)

}
\seealso{
plot.singleOut

Other Detect outliers in a time series: 
\code{\link{plot.singleOut}()}
}
\concept{Detect outliers in a time series}
