% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSpline.R
\name{estimateSplineParameters}
\alias{estimateSplineParameters}
\title{Extract estimates from fitted splines.}
\usage{
estimateSplineParameters(
  HTPSpline,
  estimate = c("predictions", "derivatives"),
  what = c("min", "max", "mean"),
  timeMin = NULL,
  timeMax = NULL,
  genotypes = NULL,
  plotIds = NULL
)
}
\arguments{
\item{HTPSpline}{An object of class HTPSpline, the output of the
\code{\link{fitSpline}} function.}

\item{estimate}{The type of estimate that should be extracted,
the predictions or the first derivatives.}

\item{what}{The type of estimate that should be extracted.}

\item{timeMin}{The lower bound of the time interval from which the
estimates should be extracted. If \code{NULL} the smallest time value for
which the splines were fitted is used.}

\item{timeMax}{The upper bound of the time interval from which the
estimates should be extracted. If \code{NULL} the largest time value for
which the splines were fitted is used.}

\item{genotypes}{A character vector indicating the genotypes for which
splines are fitted. If \code{NULL}, splines will be fitted for all genotypes.}

\item{plotIds}{A character vector indicating the plotIds for which splines
are fitted. If \code{NULL}, splines will be fitted for all plotIds.}
}
\description{
Function for extracting parameter estimates from fitted splines on a
specified interval.
}
\examples{
# Format the timepoint
spatCorrVator$timePoint <- lubridate::as_datetime(spatCorrVator$timePoint)
# Run the function to fit P-spline using the mgcv package on a subset of
# genotypes
subGeno <- c("G70","G160","G151","G179","G175","G4","G55")
fit.spline <- fitSpline(inDat = spatCorrVator,
                        trait = "EffpsII_corr",
                        genotypes = subGeno,
                        knots = 50,
                        perMinTP = 0.8)

# Estimate the maximum value of the trait at the beginning of the time course
paramVator1 <-
     estimateSplineParameters(HTPSpline = fit.spline,
                              estimate = "predictions",
                              what = "max",
                              timeMin = 1527784620,
                              timeMax = 1528500000,
                              genotypes = subGeno)

}
\seealso{
Other Fit splines: 
\code{\link{fitSpline}()},
\code{\link{plot.HTPSpline}()}
}
\concept{Fit splines}
