% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModels.R
\name{fitModels}
\alias{fitModels}
\title{Fit spatial models per time point}
\usage{
fitModels(TP, trait, covariates = NULL, geno.decomp = NULL,
  useCheck = FALSE)
}
\arguments{
\item{TP}{An object of class TP.}

\item{trait}{A character string indicating the trait used as response
variable in the model.}

\item{covariates}{A character vector indicating the variables used as
fixed effects in the model.}

\item{geno.decomp}{A character vector indicating the variables used to
group the genotypes in the model.}

\item{useCheck}{Should check genotypes be used as an extra factor in the
model?}
}
\value{
An object of class fitMod, a list of fitted spatial models.
}
\description{
Fit spatial models per time point in an object of class TP.
}
