% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSplineHDM.R
\name{fitSplineHDM}
\alias{fitSplineHDM}
\title{fitSplineHDM}
\usage{
fitSplineHDM(
  response,
  time,
  pop,
  geno,
  plant,
  data,
  dif.var = list(geno = FALSE, plant = FALSE),
  smooth.pop = list(nseg = 10, bdeg = 3, pord = 2),
  smooth.geno = list(nseg = 10, bdeg = 3, pord = 2),
  smooth.plant = list(nseg = 10, bdeg = 3, pord = 2),
  offset = NULL,
  weights = NULL,
  family = gaussian(),
  maxit = 200,
  trace = TRUE,
  thr = 0.001
)
}
\description{
fitSplineHDM
}
