% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMethods.R
\name{getBLUPsCol}
\alias{getBLUPsCol}
\title{Extract column BLUPs}
\usage{
getBLUPsCol(fitMod, outFile = NULL)
}
\arguments{
\item{fitMod}{An object of class fitMod.}

\item{outFile}{A character string indicating the .csv file to which the
results should be written. If \code{NULL} no file is written.}
}
\value{
A data.frame with column BLUPs per time point.
}
\description{
Extract column BLUPs from an object of class fitMod.
}
