% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMethods.R
\name{getBLUPsGeno}
\alias{getBLUPsGeno}
\title{Extract genotypic BLUPs}
\usage{
getBLUPsGeno(fitMod, outFile = NULL)
}
\arguments{
\item{fitMod}{An object of class fitMod.}

\item{outFile}{A character string indicating the .csv file to which the
results should be written. If \code{NULL} no file is written.}
}
\value{
A data.frame with genotypic BLUPs per time point.
}
\description{
Extract genotypic BLUPs from an object of class fitMod.
}
