% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMethods.R
\name{getBLUPsRow}
\alias{getBLUPsRow}
\title{Extract row BLUPs}
\usage{
getBLUPsRow(fitMod, outFile = NULL)
}
\arguments{
\item{fitMod}{An object of class fitMod.}

\item{outFile}{A character string indicating the .csv file to which the
results should be written. If \code{NULL} no file is written.}
}
\value{
A data.frame with row BLUPs per time point.
}
\description{
Extract row BLUPs from an object of class fitMod.
}
