% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMethods.R
\name{getColPred}
\alias{getColPred}
\title{Extract column predictions}
\usage{
getColPred(fitMod, outFile = NULL)
}
\arguments{
\item{fitMod}{An object of class fitMod.}

\item{outFile}{A character string indicting the .csv file to which the
results should be written. If \code{NULL} no file is written.}
}
\value{
A data.frame with column predictions per time point.
}
\description{
Extract column predictions from an object of class fitMod.
}
