% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMethods.R
\name{getCorrected}
\alias{getCorrected}
\title{Extract corrected phenotypic values}
\usage{
getCorrected(fitMod, outFile = NULL)
}
\arguments{
\item{fitMod}{An object of class fitMod.}

\item{outFile}{A character string indicating the .csv file to which the
results should be written. If \code{NULL} no file is written.}
}
\value{
A data.frame with spatially corrected values per time point.
}
\description{
Extract corrected phenotype from an object of class fitMod. After fitting a spatial model
at each time point, the raw phenotypic data is corrected by subtracting the (estimated)
sources of (environmental, design effect) which are of no interest (nuisances).
This allows keeping the data resolution at the plot/lant level.
}
