% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMethods.R
\name{getHerit}
\alias{getHerit}
\title{Extract heritabilities}
\usage{
getHerit(fitMod, timePoints = names(fitMod), outFile = NULL)
}
\arguments{
\item{fitMod}{An object of class fitMod.}

\item{timePoints}{A character or numeric vector indicating the timePoints
to be modeled. When using a character string to reference a timePoint, the
value has to be an exact match to one of the existing timePoints. When using
a number it will be matched by its number in the timePoints attribute of the
TP object.}

\item{outFile}{A character string indicating the .csv file to which the
results should be written. If \code{NULL} no file is written.}
}
\value{
A data.frame with heritabilities per time point.
}
\description{
Extract heritabilities from an object of class fitMod.
}
