% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getTimePoints}
\alias{getTimePoints}
\title{Extract time points}
\usage{
getTimePoints(x)
}
\arguments{
\item{x}{An object of class TP or fitMod}
}
\value{
A data.frame with columns timeNumber and timePoint listing the
time points in x
}
\description{
Function for extracting a data.frame with timeNumbers and timePoints from
an object of class TP or fitMod.
}
