% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{noCorrRoot}
\alias{noCorrRoot}
\title{Root data corrected for point outliers.}
\format{
A data.frame with 15,934 rows and 8 columns:
\describe{
  \item{timePoint}{Original time points, date and time}
  \item{Date}{Date}
  \item{thermalTime}{Thermal time cumulated}
  \item{Exp}{Experiment number}
  \item{genotype}{Genotypes}
  \item{Tank}{Tank in the greenhouse}
  \item{plotId}{Unique pot ID using rowcol coordinates}
  \item{rowId}{Row coordinate}
  \item{colId}{Column coordinate}
  \item{tipPos_y}{Position of the root tip in y axis}
}
}
\usage{
noCorrRoot
}
\description{
This dataset contains the corrected data obtained by removing point outliers
on the RootDat1 dataset. See the vignettes for details.
}
\keyword{datasets}
