% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFitMod.R
\name{plot.fitMod}
\alias{plot.fitMod}
\title{Plot function for class fitMod}
\usage{
\method{plot}{fitMod}(x, ..., plotType = c("rawPred", "corrPred",
  "herit", "effDim", "variance", "rowPred", "colPred", "timeLapse"),
  timePoints = names(x), genotypes = NULL, title = NULL,
  output = TRUE, outFile = NULL, outFileOpts = NULL)
}
\arguments{
\item{x}{An object of class fitMod.}

\item{...}{Extra plot options. Described per plotType in their respective
section.}

\item{plotType}{A single character string indicating which plot should be
made. See the sections below for a detailed explanation of the plots.}

\item{timePoints}{A character or numeric vector indicating the timePoints
to be plotted.}

\item{genotypes}{A character vector indicating the genotypes to be plotted.
Only used if \code{plotType} = "rawPred" or "corrPred".}

\item{title}{A character string used as title for the plot.}

\item{output}{Should the plot be output to the current device? If
\code{FALSE} only a list of ggplot objects is invisibly returned. Ignored if
\code{outFile} is specified.}

\item{outFile}{A character string indicting the .pdf file to which the
plots should be written. If \code{NULL} no file is written.}

\item{outFileOpts}{A named list of extra options for the pdf outfile, e.g.
width and height. See \code{\link[grDevices]{pdf}} for all possible options.}
}
\description{
Plotting function for objects of class fitMod.
}
