% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSpline.R
\name{plot.HTPSpline}
\alias{plot.HTPSpline}
\title{Plot the results of a fitted spline.}
\usage{
\method{plot}{HTPSpline}(x, ..., output = TRUE)
}
\arguments{
\item{x}{An object of class HTPSpline.}

\item{...}{Extra plot options. Described per plotType in their respective
section.}

\item{output}{Should the plot be output to the current device? If
\code{FALSE} only a list of ggplot objects is invisibly returned. Ignored if
\code{outFile} is specified.}
}
\description{
Plot the results of a fitted spline.
}
