% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSplineOrig.R
\name{plot.HTPSplineOrig}
\alias{plot.HTPSplineOrig}
\title{Plot the results of a fitted spline.}
\usage{
\method{plot}{HTPSplineOrig}(
  x,
  ...,
  plotType = c("predictions", "derivatives"),
  genotypes = NULL,
  plotIds = NULL,
  title = NULL,
  output = TRUE,
  outFile = NULL,
  outFileOpts = NULL
)
}
\arguments{
\item{x}{An object of class HTPSpline.}

\item{...}{Extra plot options. Described per plotType in their respective
section.}

\item{plotType}{A single character string indicating which plot should be
made. See the sections below for a detailed explanation of the plots.}

\item{genotypes}{A character vector indicating the genotypes for which
splines should be plotted.}

\item{plotIds}{A character vector indicating the plotIds for which splines
should be plotted.}

\item{title}{A character string used as title for the plot. If \code{NULL} a
default title is added to the plot depending on \code{plotType}.}

\item{output}{Should the plot be output to the current device? If
\code{FALSE} only a list of ggplot objects is invisibly returned. Ignored if
\code{outFile} is specified.}

\item{outFile}{A character string indicating the .pdf file to which the
plots should be written. If \code{NULL}, no file is written.}

\item{outFileOpts}{A named list of extra options for the pdf outfile, e.g.
width and height. See \code{\link[grDevices]{pdf}} for all possible options.}
}
\description{
Plot the results of a fitted spline.
}
\seealso{
Other Fit splines: 
\code{\link{estimateSplineParametersOrig}()},
\code{\link{estimateSplineParameters}()},
\code{\link{fitSplineOrig}()},
\code{\link{fitSpline}()},
\code{\link{plot.HTPSpline}()}
}
\concept{Fit splines}
