% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectPointOutliers.R
\name{plot.pointOutliers}
\alias{plot.pointOutliers}
\title{Plot point outliers}
\usage{
\method{plot}{pointOutliers}(x, ..., plotIds = NULL, outOnly = TRUE, output = TRUE)
}
\arguments{
\item{x}{An object of class pointOutliers.}

\item{...}{Extra plot options. Described per plotType in their respective
section.}

\item{plotIds}{A character vector of plotIds for which the outliers should be
detected. If \code{NULL}, all plotId in TP are used.}

\item{outOnly}{Should only plots containing outliers be plotted?}

\item{output}{Should the plot be output to the current device? If
\code{FALSE} only a list of ggplot objects is invisibly returned. Ignored if
\code{outFile} is specified.}
}
\description{
Graphical function to produce the modeled smoothing and detected outliers
for each curve of a dataset using a local regression
}
\examples{
## Create a TP object containing the data from the Phenovator.
PhenovatorDat1 <- PhenovatorDat1[!PhenovatorDat1$pos \%in\%
                                 c("c24r41", "c7r18", "c7r49"), ]
phenoTP <- createTimePoints(dat = PhenovatorDat1,
                            experimentName = "Phenovator",
                            genotype = "Genotype",
                            timePoint = "timepoints",
                            repId = "Replicate",
                            plotId = "pos",
                            rowNum = "y", colNum = "x",
                            addCheck = TRUE,
                            checkGenotypes = c("check1", "check2",
                                               "check3", "check4"))

## First select a subset of plants, for example here 10 plants
plantSel <- phenoTP[[1]]$plotId[1:9]
# Then run on the subset
resuVatorHTP <- detectPointOutliers(TP = phenoTP,
                                    trait = "EffpsII",
                                    plotIds = plantSel,
                                    confIntSize = 3,
                                    mylocfit = 0.1)

## We can then visualize the prediction by choosing a single plant...
plot(resuVatorHTP, plotIds = "c21r24", outOnly = FALSE)
## ...or a subset of plants.
plot(resuVatorHTP, plotIds = plantSel, outOnly = FALSE)

}
\seealso{
Other Detect point outliers: 
\code{\link{detectPointOutliers}()},
\code{\link{removePointOutliers}()}
}
\concept{Detect point outliers}
