% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.pred.psHDM.R
\name{plot.pred.psHDM}
\alias{plot.pred.psHDM}
\title{plot.pred.psHDM}
\usage{
\method{plot}{pred.psHDM}(
  object,
  geno.sub,
  geno.sub.names = NULL,
  geno.sub.order = NULL,
  xlab = NULL,
  ylab = NULL,
  my.theme = my.theme(),
  ask = TRUE,
  global.main = list(pop.tra = NULL, pop.tra.deriv1 = NULL, geno.tra = NULL,
    geno.tra.deriv1 = NULL, geno.dev = NULL, plant.tra = NULL)
)
}
\description{
plot.pred.psHDM
}
