% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectPointOutliers.R
\name{removePointOutliers}
\alias{removePointOutliers}
\title{Remove point outliers}
\usage{
removePointOutliers(
  TP,
  pointOutliers,
  trait = attr(x = pointOutliers, which = "trait")
)
}
\arguments{
\item{TP}{An object of class TP.}

\item{pointOutliers}{A data.frame with at least the columns plotId and
timePoint with values corresponding to those in TP. If a column outlier is
present, as in the output of \code{detectPointOutliers}, only plot x time
combinations for which outlier = 1 will be set to NA. If no column outlier is
present, all observations in pointOutliers will be set to NA.}

\item{trait}{The trait that should be set to NA. Can be ignored when using
the output of \code{detectPointOutliers} as input.}
}
\description{
Function for setting point outliers to NA.
}
