% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectSerieOut.R, R/detectSeriesOut.R
\name{removeserieOut}
\alias{removeserieOut}
\title{Remove time course outliers}
\usage{
removeserieOut(
  dat = NULL,
  fitSpline = NULL,
  serieOut,
  trait = attr(x = serieOut, which = "trait")
)

removeserieOut(
  dat = NULL,
  fitSpline = NULL,
  serieOut,
  trait = attr(x = serieOut, which = "trait")
)
}
\arguments{
\item{dat}{A data.frame.}

\item{fitSpline}{An object of class HTPSpline, the output of the
\code{\link{fitSpline}} function.}

\item{serieOut}{A data.frame with at least the column plotId with
values corresponding to those in dat.}

\item{trait}{The trait that should be set to NA. Can be ignored when using
the output of \code{detectserieOut} as input.}
}
\value{
Depending on the input either a data.frame or an object of class
HTPSpline for which the outliers specified in \code{serieOut} are
set to NA.

Depending on the input either a data.frame or an object of class
HTPSpline for which the outliers specified in \code{serieOut} are
set to NA.
}
\description{
Function for setting time course outliers in the data to NA.
The input can either be a data.frame, specified in \code{dat}, or the output
of the \code{fitSpline} function, specified in \code{fitSpline}. Exactly one
of these should be provided as input for the function.

Function for setting time course outliers in the data to NA.
The input can either be a data.frame, specified in \code{dat}, or the output
of the \code{fitSpline} function, specified in \code{fitSpline}. Exactly one
of these should be provided as input for the function.
}
\examples{
## Run the function to fit P-spline on a subset of genotypes.
subGenoVator <- c("G160", "G151")
fit.spline <- fitSpline(inDat = spatCorrVator,
                        trait = "EffpsII_corr",
                        genotypes = subGenoVator,
                        knots = 50,
                        perMinTP = 0.8)

## Extract the tables of predicted values and P-spline coefficients.
pred.Dat <- fit.spline$predDat
coef.Dat <- fit.spline$coefDat

## The coefficients are then used to tag suspect time courses
outVator <- detectserieOut(corrDat = spatCorrVator,
                            predDat = pred.Dat,
                            coefDat = coef.Dat,
                            trait = "EffpsII_corr",
                            genotypes = subGenoVator,
                            thrCor = 0.9,
                            thrPca = 30)

## The outliers can be removed from the dataset.
spatCorrVatorOut <- removeserieOut(dat = spatCorrVator,
                                    serieOut = outVator)

## Run the function to fit P-spline on a subset of genotypes.
subGenoVator <- c("G160", "G151")
fit.spline <- fitSpline(inDat = spatCorrVator,
                        trait = "EffpsII_corr",
                        genotypes = subGenoVator,
                        knots = 50,
                        perMinTP = 0.8)

## Extract the tables of predicted values and P-spline coefficients.
pred.Dat <- fit.spline$predDat
coef.Dat <- fit.spline$coefDat

## The coefficients are then used to tag suspect time courses
outVator <- detectserieOut(corrDat = spatCorrVator,
                           predDat = pred.Dat,
                           coefDat = coef.Dat,
                           trait = "EffpsII_corr",
                           genotypes = subGenoVator,
                           thrCor = 0.9,
                           thrPca = 30)

## The outliers can be removed from the dataset.
spatCorrVatorOut <- removeserieOut(dat = spatCorrVator,
                                    serieOut = outVator)

}
\seealso{
Other Detect time course outliers: 
\code{\link{detectserieOut}()},
\code{\link{plot.serieOut}()}

Other Detect time course outliers: 
\code{\link{detectserieOut}()},
\code{\link{plot.serieOut}()}
}
\concept{Detect time course outliers}
