% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectTimeCourseOutliers.R
\name{removeTimeCourseOutliers}
\alias{removeTimeCourseOutliers}
\title{Remove time course outliers}
\usage{
removeTimeCourseOutliers(
  dat,
  timeCourseOutliers,
  trait = attr(x = timeCourseOutliers, which = "trait")
)
}
\arguments{
\item{timeCourseOutliers}{A data.frame with at least the column plotId with
values corresponding to those in TP.}

\item{trait}{The trait that should be set to NA. Can be ignored when using
the output of \code{detectPointOutliers} as input.}

\item{TP}{A data.frame}
}
\description{
Function for time course outliers to NA.
}
