% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTimePoints.R
\name{removeTimePoints}
\alias{removeTimePoints}
\title{Remove selected time points from an object of class TP}
\usage{
removeTimePoints(TP, timePoints)
}
\arguments{
\item{TP}{An object of class TP.}

\item{timePoints}{A character or numeric vector indicating the time points
to be removed. When using a character string to reference a time point, the
value has to be an exact match to one of the existing timePoints. When using
a number it will be matched by its number ("timeNumber") in the timePoints
attribute of the TP object.}
}
\description{
Remove selected time points from an object of class TP
}
\seealso{
Other Data preparation: 
\code{\link{as.data.frame.TP}()},
\code{\link{createTimePoints}()},
\code{\link{plot.TP}()},
\code{\link{summary.TP}()}
}
\concept{Data preparation}
