% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spatCorrectedArch}
\alias{spatCorrectedArch}
\title{Maize data corrected for spatial trends.}
\format{
A data.frame with 37,038 rows and 9 columns:
\describe{
\item{timeNumber}{Time number obtained after formatting the original
dataset with the function createTP.}
\item{timePoint}{Original time point.}
\item{LeafArea_corr}{Leaf area, corrected data}
\item{LeafArea}{Leaf area from the picture, raw data}
\item{wt}{Weight factor}
\item{genotype}{Genotypes}
\item{geno.decomp}{Combination of treatment levels to decompose the
genotypic variance (see vignettes)}
\item{colId}{Column coordinate}
\item{rowId}{Row coordinate}
\item{plotId}{Unique pot ID using rowcol coordinates}
}
}
\usage{
spatCorrectedArch
}
\description{
This dataset contains the corrected data obtained by (1) removing outliers
for single observations and (2) running a spatial model on the
PhenoarchDat1 dataset. See the vignettes for details.
}
\keyword{datasets}
