% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spatPredArch}
\alias{spatPredArch}
\title{Maize data, genotypic predictions.}
\format{
A data.frame with 6,120 rows and 6 columns:
\describe{
\item{timeNumber}{Time number obtained after formatting the original
dataset with the function createTP.}
\item{timePoint}{Original time point.}
\item{geno.decomp}{Combination of treatment levels to decompose the
genotypic variance (see vignettes)}
\item{genotype}{Genotypes}
\item{predicted.values}{Biomass, predicted values}
\item{standard.errors}{Standard errors associated with the prediction}
}
}
\usage{
spatPredArch
}
\description{
This dataset contains the genotypic predictions obtained by (1) removing point
outliers and (2) running a spatial model on the PhenoarchDat1 dataset.
See the vignettes for details.
}
\keyword{datasets}
