% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statgenHTP.R
\docType{package}
\name{statgenHTP-package}
\alias{statgenHTP}
\alias{statgenHTP-package}
\title{statgenHTP: High Throughput Phenotyping (HTP) Data Analysis}
\description{
The statgenHTP package is being developed within the EPPN2020 
    project (<https://eppn2020.plant-phenotyping.eu/>) to streamline the analysis 
    of data coming from high throughput phenotyping platforms. It can be used 
    for spatial analysis of time-series data. 
    Some functions have been created to be used in conjunction with the R 
    package 'asreml' for the 'ASReml' software, which can be obtained upon 
    purchase from 'VSN' international (<https://www.vsni.co.uk/software/asreml>).
}
\author{
\strong{Maintainer}: Bart-Jan van Rossum \email{bart-jan.vanrossum@wur.nl} (\href{https://orcid.org/0000-0002-8673-2514}{ORCID})

Authors:
\itemize{
  \item Emilie J Millet \email{emilie.millet@wur.nl} (\href{https://orcid.org/0000-0002-2913-4892}{ORCID})
  \item Maria Xose Rodriguez Alvarez (\href{https://orcid.org/0000-0002-1329-9238}{ORCID})
  \item Diana Marcela Perez Valencia (\href{https://orcid.org/0000-0002-9053-2929}{ORCID})
  \item Isabelle Sanchez
  \item Nadine Hilgert
  \item Fred van Eeuwijk (\href{https://orcid.org/0000-0003-3672-2921}{ORCID})
  \item Martin Boer
}

}
\keyword{internal}
