% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFitMod.R
\name{[.fitMod}
\alias{[.fitMod}
\title{Function for extracting objects of class fitMod that keeps class.}
\usage{
\method{[}{fitMod}(x, i, ...)
}
\arguments{
\item{x}{An object of class fitMod.}

\item{i}{An index specifying the element to extract of replace.}

\item{...}{Ignored.}
}
\description{
Function for extracting objects of class fitMod that keeps class.
}
