% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTP.R
\name{[.TP}
\alias{[.TP}
\title{Function for extracting for objects of class TP that keeps class.}
\usage{
\method{[}{TP}(x, i, ...)
}
\description{
Function for extracting for objects of class TP that keeps class.
}
