% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTimePoints.R
\name{summary.TP}
\alias{summary.TP}
\title{Summary function for class TP}
\usage{
\method{summary}{TP}(object, ...)
}
\arguments{
\item{object}{An object of class TP.}

\item{...}{Ignored.}
}
\description{
Summary function for class TP
}
