% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.psHDM.R
\name{themeHDM}
\alias{themeHDM}
\title{themeHDM}
\usage{
themeHDM(textSize = 20)
}
\arguments{
\item{textSize}{Numeric value, font size of the texts in the plots.}
}
\description{
Custom theme for creating HDM plots.
}
