% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctSpatial.R
\name{correctSpatial}
\alias{correctSpatial}
\title{Correct for spatial effects and other unnecessary factors.}
\usage{
correctSpatial(fitMod)
}
\description{
Correct for spatial effects and other unnecessary factors.
}
\keyword{internal}
