% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{countValid}
\alias{countValid}
\title{Count valid observations per time point for a given trait}
\usage{
countValid(TP, trait)
}
\arguments{
\item{TP}{An object of class TP.}

\item{trait}{A character string indicating the trait for which valid
observations should be counted.}
}
\description{
Count valid observations per time point for a given trait.
}
