% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTimePoints.R
\name{createTimePoints}
\alias{createTimePoints}
\title{Create an object of class TP}
\usage{
createTimePoints(dat, genotype, timePoint, plotId, repId = NULL,
  rowNum = NULL, colNum = NULL, addCheck = FALSE,
  checkGenotypes = NULL)
}
\arguments{
\item{dat}{A data.frame.}

\item{genotype}{A character string indicating the column in dat containing
the genotypes.}

\item{timePoint}{A character string indicating the column in dat containing
the time points.}

\item{plotId}{A character string indicating the column in dat containing
the plotId. This has to be a unique identifier per plot/plant.}

\item{repId}{A character string indicating the column in dat containing
the replicates.}

\item{rowNum}{A character string indicating the column in dat containing
the row number of the plot.}

\item{colNum}{A character string indicating the column in dat containing
the column number of the plot.}

\item{addCheck}{Should a column check be added to the output? If \code{TRUE},
checkGenotypes cannot be \code{NULL}.}

\item{checkGenotypes}{A character vector containing the genotypes used as
checks in the experiment.}
}
\value{
An object of class TP. A list with, per time point in the input, a
data.frame containing the data for that time point. A data.frame with
columns timeNumber and timePoint is added as attribute timePoints to the
data. This data.frame can be used for referencing timePoints by their
number.
}
\description{
Convert a data.frame to an object of class TP (Time Points).
The function converts a data.frame to an object of class TP in the following
steps:
\itemize{
\item{Check input data}
\item{Rename columns to default column names - default column names:
genotype, timePoint, plotId, repId, rowNum, colNum}.
\item{Add columns rowId and colId to data. These are copies of rowId and
colId converted to factors.}
\item{Convert column types to default column types - rowNum and colNum
are converted to numeric columns, all other renamed columns to factor
columns. Columns other than the default columns, e.g. traits or other
covariates will be included in the output unchanged.}
\item{If \code{addCheck = TRUE}, add a column check with a value
"noCheck" for the genotypes that are not in \code{checkGenotypes} and
the name of the genotype for the \code{checkGenotypes}. Also a column
genoCheck is added with the names of the genotypes that are not in
\code{checkGenotypes} and \code{NA} for the \code{checkGenotypes}. These
columns are necessary for fitting models on data that includes check
genotypes, e.g. reference genotypes that are highly replicated or in case
of augmented design.}
\item{Split input data by time point - each time point in the input data will
become a list item in the output.}
\item{Add a list of time points as attribute \code{timePoints} to the
output.}
}
Note that \code{plotId} needs to be a unique identifier for a plot or a
plant. It cannot occur more than once per time point.
}
\seealso{
Other functions for TP objects: \code{\link{plot.TP}}
}
\concept{functions for TP objects}
