% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectPointOutliers.R
\name{detectPointOutliers}
\alias{detectPointOutliers}
\title{detectPointOutliers}
\usage{
detectPointOutliers(TP, trait, plotIds = NULL, confIntSize = 5, mylocfit = 0.5)
}
\arguments{
\item{trait}{character, name of the variable to model in datain (for example,
Biomass, PH or LA and so on)}

\item{plotIds}{A character vector of plotIds for which the outliers should be
detected. If \code{NULL}}

\item{confIntSize}{numeric, factor to calculate the confidence interval}

\item{mylocfit}{numeric, The constant component of the smoothing parameter.
(see the locfit())}

\item{datain}{input dataframe. This dataframe contains a set of time courses}
}
\value{
An object of class pointOutliers, a data.frame with the following
columns.
\describe{
  \item{plotId}{the id variable}
  \item{timePoint}{time variable in datain}
  \item{trait}{name of the modeled variable in datain}
  \item{yPred}{the locfit prediction}
  \item{sd_yPred}{standard deviation of the prediction}
  \item{lwr}{lower bound of the confidence interval}
  \item{upr}{upper bound of the confidence interval}
  \item{outlier}{flag of detected outlier (0 is outlier, 1 is not)}
}
}
\description{
Function to model each curve of a dataset using a local regression
}
\details{
see locfit() help function from the locfit R library. The user can act on:
\describe{
  \item{mylocfit}{the constant of the smoothing parameter. Increase mylocfit
  to have a very smooth curve}
  \item{confIntSize}{the level to calculate the confidence interval. Increase
  confIntSize to exclude less outliers}
}
to produce the grahics of the prediction and detected outliers, please use
plotDetectPointOutlierLocFit() function.
}
\seealso{
plot.pointOutliers
}
