% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectSerieOut.R, R/detectSeriesOut.R
\name{detectserieOut}
\alias{detectserieOut}
\title{Outlier detection}
\usage{
detectserieOut(
  corrDat,
  predDat,
  coefDat,
  trait,
  genotypes = NULL,
  geno.decomp = NULL,
  thrCor = 0.9,
  thrPca = 30
)

detectserieOut(
  corrDat,
  predDat,
  coefDat,
  trait,
  genotypes = NULL,
  geno.decomp = NULL,
  thrCor = 0.9,
  thrPca = 30
)
}
\arguments{
\item{corrDat}{A data.frame with corrected spatial data.}

\item{predDat}{A data.frame with predicted data from a fitted spline.}

\item{coefDat}{A data.frame with coefficients from a fitted spline.}

\item{trait}{A character string indicating the trait for which to detect
outliers.}

\item{genotypes}{A character vector indicating the genotypes for which to
detect outliers. If \code{NULL}, outlier detection will be done for all
genotypes.}

\item{geno.decomp}{A character vector indicating the variables to use to
group the genotypic variance in the model.}

\item{thrCor}{A numerical value used as threshold for determining outliers
based on correlation between plots.}

\item{thrPca}{A numerical value used as threshold for determining outliers
based on angles (in degrees) between PCA scores.}
}
\description{
Function for detecting strange time courses. The function uses the estimates
for the spline coefficients per time course (typically per plant).
Correlations between those coefficient vectors are calculated to identify
outlying time courses, i.e., plants. An outlying time course will have low
correlation to the majority of time courses. To support the analysis by
correlations, a principal component analysis can be done on the plant
(time course) by spline coefficient matrix. A PCA plot of the plant scores
will show the outlying plants.

Function for detecting strange time courses. The function uses the estimates
for the spline coefficients per time course (typically per plant).
Correlations between those coefficient vectors are calculated to identify
outlying time courses, i.e., plants. An outlying time course will have low
correlation to the majority of time courses. To support the analysis by
correlations, a principal component analysis can be done on the plant
(time course) by spline coefficient matrix. A PCA plot of the plant scores
will show the outlying plants.
}
\examples{
\donttest{
## The data from the Phenovator platform have been corrected for
## spatial trends and time points outliers have been removed.

# Run the function to fit P-splines on a subset of
## genotypes
subGenoVator <- c("G160", "G151")
fit.spline <- fitSpline(inDat = spatCorrVator,
                        trait = "EffpsII_corr",
                        genotypes = subGenoVator,
                        knots = 50,
                        perMinTP = 0.8)

## Extract the tables of predicted values and P-spline coefficients
pred.Dat <- fit.spline$predDat
coef.Dat <- fit.spline$coefDat

## The coefficients are then used to tag suspect time courses.
outVator <- detectserieOut(corrDat = spatCorrVator,
                            predDat = pred.Dat,
                            coefDat = coef.Dat,
                            trait = "EffpsII_corr",
                            genotypes = subGenoVator,
                            thrCor = 0.9,
                            thrPca = 30)

## The `outVator` can be visualized for selected genotypes.
plot(outVator, genotypes = "G151")
}

\donttest{
## The data from the Phenovator platform have been corrected for
## spatial trends and time points outliers have been removed.

# Run the function to fit P-splines on a subset of
## genotypes
subGenoVator <- c("G160", "G151")
fit.spline <- fitSpline(inDat = spatCorrVator,
                        trait = "EffpsII_corr",
                        genotypes = subGenoVator,
                        knots = 50,
                        perMinTP = 0.8)

## Extract the tables of predicted values and P-spline coefficients
pred.Dat <- fit.spline$predDat
coef.Dat <- fit.spline$coefDat

## The coefficients are then used to tag suspect time courses.
outVator <- detectserieOut(corrDat = spatCorrVator,
                           predDat = pred.Dat,
                           coefDat = coef.Dat,
                           trait = "EffpsII_corr",
                           genotypes = subGenoVator,
                           thrCor = 0.9,
                           thrPca = 30)

## The `outVator` can be visualized for selected genotypes.
plot(outVator, genotypes = "G151")
}

}
\seealso{
Other Detect time course outliers: 
\code{\link{plot.serieOut}()},
\code{\link{removeserieOut}()}

Other Detect time course outliers: 
\code{\link{plot.serieOut}()},
\code{\link{removeserieOut}()}
}
\concept{Detect time course outliers}
