% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectTimeCourseOutliers.R
\name{detectTimeCourseOutliers}
\alias{detectTimeCourseOutliers}
\title{Outlier detection}
\usage{
detectTimeCourseOutliers(
  corrDat,
  predDat,
  coefDat,
  trait,
  genotypes = NULL,
  geno.decomp = NULL,
  thrCor = 0.9,
  thrPca = 1
)
}
\arguments{
\item{corrDat}{A data.frame with corrected spatial data.}

\item{predDat}{A data.frame with predicted data from a fitted spline.}

\item{coefDat}{A data.frame with coefficients from a fitted spline.}

\item{trait}{A character string indicating the trait for which to detect
outliers.}

\item{genotypes}{A character vector indicating the genotypes for which to
detect outliers. If \code{NULL}, outlier detection will be done for all
genotypes.}

\item{geno.decomp}{A character vector indicating the variables to use to
group the genotypic variance in the model.}

\item{thrCor}{A numerical value used as threshold for determining outliers
based on correlation between plots.}

\item{thrPca}{A numerical value used as threshold for determining outliers
based on PCA scores.}
}
\description{
Function for detecting outliers.
}
\examples{
## The data from the Phenovator platform have been corrected for
## spatial trends and time points outliers have been removed.
# Format the timepoint
spatCorrVator$timePoint <- lubridate::as_datetime(spatCorrVator$timePoint)
# Run the function to fit p-spline using the mgcv package on a subset of
# genotypes
subGenoVator <- c("G70","G160","G151","G179","G175","G4","G55")
fit.spline <- fitSpline(inDat = spatCorrVator,
                        trait = "EffpsII_corr",
                        genotypes = subGenoVator,
                        knots = 50,
                        perMinTP = 0.8)
# Extracting the tables of predicted values and pspline coeficients
pred.Dat <- fit.spline$predDat
coef.Dat <- fit.spline$coefDat

## The coefficients are then used to tag suspect time courses
outVator <- detectTimeCourseOutliers(corrDat = spatCorrVator,
                         predDat = pred.Dat,
                         coefDat = coef.Dat,
                         trait = "EffpsII_corr",
                         genotypes = subGenoVator,
                         thrCor = 0.9,
                         thrPca = 1)
# The `outVator` can be visualized by selecting genotypes
plot(outVator, genotypes = "G151")

}
\seealso{
Other Detect time course outliers: 
\code{\link{plot.timeCourseOutliers}()},
\code{\link{removeTimeCourseOutliers}()}
}
\concept{Detect time course outliers}
