% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSpline.R
\name{fitSpline}
\alias{fitSpline}
\title{Fit Splines}
\usage{
fitSpline(
  inDat,
  trait,
  genotypes = NULL,
  plotIds = NULL,
  knots = 50,
  useTimeNumber = FALSE,
  timeNumber = NULL,
  minNoTP = NULL
)
}
\arguments{
\item{inDat}{A data.frame with corrected spatial data.}

\item{trait}{A character string indicating the trait for which the spline
should be fitted.}

\item{genotypes}{A character vector indicating the genotypes for which
splines are fitted. If \code{NULL}, splines will be fitted for all genotypes.}

\item{plotIds}{A character vector indicating the plotIds for which splines
are fitted. If \code{NULL}, splines will be fitted for all plotIds.}

\item{knots}{The number of knots to use when fitting the spline.}

\item{useTimeNumber}{Should the timeNumber be used instead of the timePoint?}

\item{timeNumber}{If \code{useTimeNumber = TRUE}, a character vector
indicating the column containing the numerical time to use.}

\item{minNoTP}{The minimum number of time point for which data should be
available for a plot. Defaults to 80\% of all time points present in the
TP object.}
}
\description{
Fit P-Splines on corrected or raw data. The number of
knots are chosen by the user. The function outputs are predicted P-Spline
values and its first derivative at a denser time step. The idea is to use
the outputs for outlier detection and to estimate relevant parameters from
the curve.
}
\examples{
## The data from the Phenovator platform have been corrected for
## spatial trends and time points outliers have been removed.

## Run the function to fit P-Spline on a subset of
## genotypes
subGeno <- c("G070", "G160")
fit.spline <- fitSpline(inDat = spatCorrectedVator,
                        trait = "EffpsII_corr",
                        genotypes = subGeno,
                        knots = 50)

## Extract the tables of predicted values and P-Spline coefficients.
pred.Dat <- fit.spline$predDat
coef.Dat <- fit.spline$coefDat

## We can then visualize the P-Spline predictions and first derivatives
## for a subset of genotypes...
plot(fit.spline, genotypes = "G160")

## ...or for a subset of plots.
plot(fit.spline, plotIds = "c10r29", plotType =  "predictions")
plot(fit.spline, plotIds = "c10r29", plotType =  "derivatives")

}
\seealso{
Other Fit splines: 
\code{\link{estimateSplineParameters}()},
\code{\link{plot.HTPSpline}()}
}
\concept{Fit splines}
