% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSpline.R
\name{fitSpline}
\alias{fitSpline}
\title{Fit Splines}
\usage{
fitSpline(
  corrDat,
  trait,
  genotypes = NULL,
  plotIds = NULL,
  knots = 50,
  useTimeNumber = FALSE,
  timeNumber = NULL,
  perMinTP = 0.8
)
}
\arguments{
\item{corrDat}{A data.frame with corrected spatial data.}

\item{trait}{A character string indicating the trait for which the spline
should be fitted.}

\item{genotypes}{A character vector indicating the genotypes for which
splines are fitted. If \code{NULL}, splines will be fitted for all genotypes.}

\item{plotIds}{A character vector indicating the plotIds for which splines
are fitted. If \code{NULL}, splines will be fitted for all plotIds.}

\item{knots}{The number of knots to use when fitting the spline.}

\item{useTimeNumber}{Should the timeNumber be used instead of the timePoint?}

\item{timeNumber}{If \code{useTimeNumber = TRUE}, a character vector
indicating the column containing the numerical time to use.}

\item{perMinTP}{The percentage of minimum number of time points to use.}
}
\description{
Function for fitting splines.
}
\examples{
## The data from the Phenovator platform have been corrected for
# spatial trends and time points outliers have been removed
# Format the timepoint
spatCorrVator$timePoint <- lubridate::as_datetime(spatCorrVator$timePoint)
# Run the function to fit p-spline using the mgcv package on a subset of
# genotypes
subGeno <- c("G70","G160","G151","G179","G175","G4","G55")
fit.spline <- fitSpline(corrDat = spatCorrVator,
                        trait = "EffpsII_corr",
                        genotypes = subGeno,
                        knots = 50,
                        perMinTP = 0.8)
# Extracting the tables of predicted values and pspline coeficients
pred.Dat <- fit.spline$predDat
coef.Dat <- fit.spline$coefDat

## We can then visualise the p-spline predictions and first derivatives
# for a subset of genotypes...
plot(fit.spline, genotypes = "G160")
# ...or for a subset of plots.
plot(fit.spline, plotIds = "c10r29", plotType =  "predictions")
plot(fit.spline, plotIds = "c10r29", plotType =  "derivatives")

}
\seealso{
Other Fit splines: 
\code{\link{estimateSplineParameters}()},
\code{\link{plot.HTPSpline}()}
}
\concept{Fit splines}
