% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSplineHDM.R
\name{fitSplineHDM}
\alias{fitSplineHDM}
\title{fitSplineHDM}
\usage{
fitSplineHDM(
  inDat,
  trait,
  time,
  pop,
  geno,
  plant,
  dif.var = list(geno = FALSE, plant = FALSE),
  smooth.pop = list(nseg = 10, bdeg = 3, pord = 2),
  smooth.geno = list(nseg = 10, bdeg = 3, pord = 2),
  smooth.plant = list(nseg = 10, bdeg = 3, pord = 2),
  offset = NULL,
  weights = NULL,
  family = gaussian(),
  maxit = 200,
  trace = TRUE,
  thr = 0.001
)
}
\arguments{
\item{inDat}{A data.frame with corrected spatial data.}

\item{trait}{A character string indicating the trait for which the spline
should be fitted.}
}
\description{
fitSplineHDM
}
