% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMethods.R
\name{getEffDims}
\alias{getEffDims}
\title{Extract effective dimensions}
\usage{
getEffDims(fitMod, outFile = NULL)
}
\arguments{
\item{fitMod}{An object of class fitMod.}

\item{outFile}{A character string indicating the .csv file to which the
results should be written. If \code{NULL} no file is written.}
}
\value{
A data.frame with effective dimensions per time point.
}
\description{
Extract effective dimensions from an object of class fitMod.
}
