% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMethods.R
\name{getHerit}
\alias{getHerit}
\title{Extract heritabilities}
\usage{
getHerit(fitMod, outFile = NULL)
}
\arguments{
\item{fitMod}{An object of class fitMod.}

\item{outFile}{A character string indicting the .csv file to which the
results should be written. If \code{NULL} no file is written.}
}
\value{
A data.frame with heritabilities per time point.
}
\description{
Extract heritabilities from an object of class fitMod.
}
