% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFitMod.R
\name{plot.fitMod}
\alias{plot.fitMod}
\title{Plot function for class fitMod}
\usage{
\method{plot}{fitMod}(x, ..., plotType = c("rawPred", "corrPred",
  "herit", "effDim", "variance", "timeLapse"), whichED = c("colId",
  "rowId", "fCol", "fRow", "fColRow", "colfRow", "fColfRow", "surface"),
  timePoints = names(x), genotypes = NULL, title = NULL,
  output = TRUE, outFile = NULL, outFileOpts = NULL)
}
\arguments{
\item{x}{An object of class fitMod.}

\item{...}{Extra plot options. Described per plotType in their respective
section.}

\item{plotType}{A single character string indicating which plot should be
made. See the sections below for a detailed explanation of the plots.}

\item{whichED}{A character vector indicating which effective dimensions
shoul be plotted. Only used if \code{plotType} = "effDim".}

\item{timePoints}{A character or numeric vector indicating the timePoints
to be plotted. When using a character string to reference a timePoint, the
value has to be an exact match to one of the existing timePoints. When using
a number it will be matched by its number in the timePoints attribute of the
TP object.}

\item{genotypes}{A character vector indicating the genotypes to be plotted.
Only used if \code{plotType} = "rawPred" or "corrPred".}

\item{title}{A character string used as title for the plot. If \code{NULL} a
default title is added to the plot depending on \code{plotType}.}

\item{output}{Should the plot be output to the current device? If
\code{FALSE} only a list of ggplot objects is invisibly returned. Ignored if
\code{outFile} is specified.}

\item{outFile}{A character string indicating the .pdf file or .gif file
(For \code{plotType} = "timeLapse") to which the plots should be written.}

\item{outFileOpts}{A named list of extra options for the pdf outfile, e.g.
width and height. See \code{\link[grDevices]{pdf}} for all possible options.}
}
\value{
Depending on the plot type either a ggplot object or a list of ggplot
objects is invisibly returned.
}
\description{
Plotting function for objects of class fitMod.
}
\section{rawPred plot}{

Plots the raw data overlayed with the predicted values from the fitted model.
For each genotype a plot is made per plot/plant over time. These plots are
put together in a 5x5 grid. By using the parameter \code{genotypes} a
selection of genotypes can be plotted.
}

\section{corrPred plot}{

Plots the spatially corrected data overlayed with the predicted values from
the fitted model. For each genotype a plot is made per plot/plant over time.
These plots are put together in a 5x5 grid. By using the parameter
\code{genotypes} a selection of genotypes can be plotted.
}

\section{herit plot}{

Plots the heritability over time. If \code{geno.decomp} is used when fitting
the model, heritabilities are plotted for each level of geno.decomp in a
single plot.
}

\section{effDim plot}{

Plots the effective dimension from models fitted using SpATS over time.
}

\section{variance plot}{

Plots the residual, column and row variance for the fitted model over time.
}

\section{timeLapse plot}{

Creates a time lapse of the spatial trends of models fitted using SpATS over
time.
}

