% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectPointOutliers.R
\name{plot.pointOutliers}
\alias{plot.pointOutliers}
\title{plotDetectPointOutlierLocFit}
\usage{
\method{plot}{pointOutliers}(x, ..., plotIds = NULL, outOnly = TRUE, output = TRUE)
}
\arguments{
\item{x}{An object of class pointOutliers.}

\item{...}{Extra plot options. Described per plotType in their respective
section.}

\item{plotIds}{A character vector of plotIds for which the outliers should be
detected. If \code{NULL}, all plotId in TP are used.}

\item{outOnly}{Should only plots containing outliers be plotted?}

\item{output}{Should the plot be output to the current device? If
\code{FALSE} only a list of ggplot objects is invisibly returned. Ignored if
\code{outFile} is specified.}
}
\description{
graphical function to produce the modeled smoothing and detected outliers
for each curve of a dataset using a local regression
}
