% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectTimeCourseOutliers.R
\name{plot.timeCourseOutliers}
\alias{plot.timeCourseOutliers}
\title{plot.timeCourseOutliers}
\usage{
\method{plot}{timeCourseOutliers}(x, ..., genotypes = NULL, title = NULL, output = TRUE)
}
\arguments{
\item{x}{An object of class timeCourseOutliers.}

\item{...}{Extra plot options. Described per plotType in their respective
section.}

\item{genotypes}{A character vector indicating the genotypes for which to
detect outliers. If \code{NULL}, outlier detection will be done for all
genotypes.}

\item{title}{A character string used as title for the plot. If \code{NULL} a
default title is added to the plot depending on \code{plotType}.}

\item{output}{Should the plot be output to the current device? If
\code{FALSE} only a list of ggplot objects is invisibly returned. Ignored if
\code{outFile} is specified.}
}
\description{
plot.timeCourseOutliers
}
\examples{
## The data from the Phenovator platform have been corrected for spatial
## trends and time points outliers have been removed.
# Format the timepoint
spatCorrVator$timePoint <- lubridate::as_datetime(spatCorrVator$timePoint)
# Run the function to fit p-spline using the mgcv package on a subset of
# genotypes
subGeno <- c("G70","G160","G151","G179","G175","G4","G55")
fit.spline <- fitSpline(corrDat = spatCorrVator,
                        trait = "EffpsII_corr",
                        genotypes = subGeno,
                        knots = 50,
                        perMinTP = 0.8)
# Extracting the tables of predicted values and pspline coeficients
pred.Dat <- fit.spline$predDat
coef.Dat <- fit.spline$coefDat

## The coefficients are then used to tag suspect time courses
outVator <- detectTimeCourseOutliers(corrDat = spatCorrVator,
                         predDat = pred.Dat,
                         coefDat = coef.Dat,
                         trait = "EffpsII_corr",
                         genotypes = subGeno,
                         thrCor = 0.9,
                         thrPca = 1)
# The `outVator` can be visualised by selecting genotypes
plot(outVator, genotypes = "G151")

}
