% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTimePoints.R
\name{plot.TP}
\alias{plot.TP}
\title{Plot function for class TP}
\usage{
\method{plot}{TP}(x, ..., plotType = c("layout", "box", "cor", "raw"),
  timePoints = names(x), traits = NULL, genotypes = NULL,
  geno.decomp = NULL, output = TRUE)
}
\arguments{
\item{x}{An object of class TP.}

\item{...}{Extra plot options. Described per plotType in their respective
section.}

\item{plotType}{A single character string indicating which plot should be
made. See the sections below for a detailed explanation of the plots.}

\item{timePoints}{A character vector indicating the timePoints to be plotted.
Only used if \code{plotType} = "layout" or "box".}

\item{traits}{A character vector indicating the traits to be plotted in
a boxplot. Only used if \code{plotType} = "box" or "cor".}

\item{genotypes}{A character vector indicating the genotypes to be plotted.
Only used if \code{plotType} = "raw".}

\item{geno.decomp}{A character vector indicating the grouping of the
genotypes to be plotted. Only used if \code{plotType} = "raw".}

\item{output}{Should the plot be output to the current device? If
\code{FALSE} only a list of ggplot objects is invisibly returned.}
}
\description{
Plotting function for objects of class TP. Plots either the layout of the
different timePoints within the TP object or locates the timePoints on a map. Also a
boxplot can be made for selected traits and timePoints and a plot of
correlations
between timePoints. A detailed description and optional extra parameters of the
different plots is given in the sections below.
}
\section{Layout Plot}{

Plots the layout of the selected timePoints (all available timePoints by default).
This plot can only be made for timePoints that contain both row (\code{rowNum})
and column (\code{colNum}) information. If either one of those is missing
the timePoint is skipped with a warning. If blocks (\code{subBlock}) are
available for a timePoint these are indicated in different colors per block,
otherwise all plots are colored in grey. If replicates (\code{repId}) are
available a black line is plotted between diffent replicates. Missing plots
are indicated in white. This can either be single plots in a timePoint or
complete missing columns or rows.\cr
Extra parameter options:
\describe{
\item{showGeno}{Should individual genotypes be indicated in the plot?
Defaults to \code{FALSE}}
\item{highlight}{A character vector of genotypes to be highlighted in the
plot.}
\item{colorSubBlock}{Should subBlocks be colored with a different color per
subBlock? Defaults to \code{FALSE}. \code{colorSubBlock} is ignored when
highlight is used to highlight genotypes.}
}
}

\section{Box Plot}{

Creates a boxplot per selected trait grouped by timePoint. Extra parameter
options:
\describe{
\item{groupBy}{A character string indicating a column in \code{TP} by which
the boxes in the plot should be grouped. By default the boxes are grouped
per timePoint.}
\item{colorBy}{A character string indicating a column in \code{TP} by which
the boxes are colored. Coloring will be done within the groups indicated by
the \code{groupBy} parameter.}
\item{orderBy}{A character string indicating the way the boxes should be
ordered. Either "alphabetic" for alphabetical ordering of the groups,
"ascending" for ordering by ascending mean, or "descending" for ordering by
descending mean. Default boxes are ordered alphabetically.}
}
}

\section{Correlation Plot}{

Draws a heatmap of correlations between timePoints per selected trait. If
genotypes are replicated within timePoints genotypic means are taken before
computing correlations. The order of the timePoints in the heatmap is determined
by clustering them.
}

\concept{functions for TP objects}
