% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.psHDM.R
\name{predict.psHDM}
\alias{predict.psHDM}
\title{predict.psHDM}
\usage{
\method{predict}{psHDM}(
  object,
  newtimes,
  pred = list(pop = TRUE, geno = TRUE, plant = TRUE),
  se = list(pop = TRUE, geno = TRUE, plant = FALSE),
  ...
)
}
\description{
predict.psHDM
}
