% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectPointOutliers.R
\name{removePointOutliers}
\alias{removePointOutliers}
\title{Remove point outliers}
\usage{
removePointOutliers(
  TP,
  pointOutliers,
  trait = attr(x = pointOutliers, which = "trait")
)
}
\arguments{
\item{TP}{An object of class TP.}

\item{pointOutliers}{A data.frame with at least the columns plotId and
timePoint with values corresponding to those in TP. If a column outlier is
present, as in the output of \code{detectPointOutliers}, only plot x time
combinations for which outlier = 1 will be set to NA. If no column outlier is
present, all observations in pointOutliers will be set to NA.}

\item{trait}{The trait that should be set to NA. Can be ignored when using
the output of \code{detectPointOutliers} as input.}
}
\description{
Function for setting point outliers to NA.
}
\examples{
## Create a TP object containing the data from the Phenovator.
PhenovatorDat1 <- PhenovatorDat1[!PhenovatorDat1$pos \%in\% c("c24r41", "c7r18","c7r49"),]
phenoTP <- createTimePoints(dat = PhenovatorDat1,
                            experimentName = "Phenovator",
                            genotype = "Genotype",
                            timePoint = "timepoints",
                            repId = "Replicate",
                            plotId = "pos",
                            rowNum = "y", colNum = "x",
                            addCheck = TRUE,
                            checkGenotypes = c("check1","check2","check3","check4"))

## First select a subset of plants, for example here 10 plants
plantSel <- phenoTP[[1]]$plotId[1:9]
# plantSel <- c("c11r55","c6r49","c21r16","c13r45")
# Then run on the subset
resuVatorHTP <- detectPointOutliers(TP = phenoTP,
                                    trait = "EffpsII",
                                    plotIds = plantSel,
                                    confIntSize = 3,
                                    mylocfit = 0.1)

## The annotated points can be replaced by NA for the studied trait
phenoTPOut <- removePointOutliers(phenoTP, resuVatorHTP)

}
