% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectTimeCourseOutliers.R
\name{removeTimeCourseOutliers}
\alias{removeTimeCourseOutliers}
\title{Remove time course outliers}
\usage{
removeTimeCourseOutliers(
  dat = NULL,
  fitSpline = NULL,
  timeCourseOutliers,
  trait = attr(x = timeCourseOutliers, which = "trait")
)
}
\arguments{
\item{dat}{A data.frame.}

\item{fitSpline}{An object of class HTPSpline, the output of the
\code{\link{fitSpline}} function.}

\item{timeCourseOutliers}{A data.frame with at least the column plotId with
values corresponding to those in dat.}

\item{trait}{The trait that should be set to NA. Can be ignored when using
the output of \code{timeCourseOutliers} as input.}
}
\value{
Depending on the input either a data.frame or an object of class
HTPSpline for which the outliers specified in \code{timeCourseOutliers} are
set to NA.
}
\description{
Function for setting time course outliers in the data to NA.
The input can either be a data.frame, specified in \code{dat}, or the output
of the fitSpline function, specified in \code{fitSpline}. Exactly one of
these should be provided as input for the function.
}
\examples{
# Format the timepoint
spatCorrVator$timePoint <- lubridate::as_datetime(spatCorrVator$timePoint)
# Run the function to fit p-spline using the mgcv package on a subset of
# genotypes
subGenoVator <- c("G70","G160","G151","G179","G175","G4","G55")
fit.spline <- fitSpline(inDat = spatCorrVator,
                        trait = "EffpsII_corr",
                        genotypes = subGenoVator,
                        knots = 50,
                        perMinTP = 0.8)
# Extracting the tables of predicted values and pspline coeficients
pred.Dat <- fit.spline$predDat
coef.Dat <- fit.spline$coefDat

## The coefficients are then used to tag suspect time courses
outVator <- detectTimeCourseOutliers(corrDat = spatCorrVator,
                         predDat = pred.Dat,
                         coefDat = coef.Dat,
                         trait = "EffpsII_corr",
                         genotypes = subGenoVator,
                         thrCor = 0.9,
                         thrPca = 1)

## The outliers can be removed from the dataset
spatCorrVatorOut <-
       removeTimeCourseOutliers(dat = spatCorrVator,
                                timeCourseOutliers = outVator)

}
\seealso{
Other Detect time course outliers: 
\code{\link{detectTimeCourseOutliers}()},
\code{\link{plot.timeCourseOutliers}()}
}
\concept{Detect time course outliers}
