% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statgenHTP.R
\docType{package}
\name{statgenHTP-package}
\alias{statgenHTP}
\alias{statgenHTP-package}
\title{statgenHTP: High Throughput Phenotyping (HTP) Data Analysis}
\description{
Phenotypic analysis of data coming from high throughput phenotyping (HTP) platforms, including different types of outlier detection, spatial analysis, and parameter estimation. The package is being developed within the EPPN2020 project (\url{https://eppn2020.plant-phenotyping.eu/}). Some functions have been created to be used in conjunction with the R package 'asreml' for the 'ASReml' software, which can be obtained upon purchase from 'VSN' international (\url{https://www.vsni.co.uk/software/asreml}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://biometris.github.io/statgenHTP/index.html}
  \item \url{https://github.com/Biometris/statgenHTP/}
  \item Report bugs at \url{https://github.com/Biometris/statgenHTP/issues}
}

}
\author{
\strong{Maintainer}: Bart-Jan van Rossum \email{bart-jan.vanrossum@wur.nl} (\href{https://orcid.org/0000-0002-8673-2514}{ORCID})

Authors:
\itemize{
  \item Emilie J Millet \email{emilie.millet@wur.nl} (\href{https://orcid.org/0000-0002-2913-4892}{ORCID})
  \item Maria Xose Rodriguez Alvarez (\href{https://orcid.org/0000-0002-1329-9238}{ORCID})
  \item Diana Marcela Perez Valencia (\href{https://orcid.org/0000-0002-9053-2929}{ORCID})
  \item Isabelle Sanchez
  \item Nadine Hilgert
  \item Fred van Eeuwijk (\href{https://orcid.org/0000-0003-3672-2921}{ORCID})
  \item Martin Boer
}

}
\keyword{internal}
