% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTimePoints.R
\name{[.TP}
\alias{[.TP}
\title{Function for extracting for objects of class TP that keeps class and
attributes.}
\usage{
\method{[}{TP}(x, i, ...)
}
\arguments{
\item{x}{An object of class TP.}

\item{i}{An index specifying the element to extract of replace.}

\item{...}{Ignored.}
}
\description{
Function for extracting for objects of class TP that keeps class and
attributes.
}
