% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFitMod.R
\name{summary.fitMod}
\alias{summary.fitMod}
\title{Summary function for class fitMod}
\usage{
\method{summary}{fitMod}(object, ...)
}
\arguments{
\item{object}{An object of class fitMod.}

\item{...}{Ignored.}
}
\description{
Summary function for class fitMod
}
\seealso{
Other Spatial modeling: 
\code{\link{fitModels}()},
\code{\link{getCorrected}()},
\code{\link{getEffDims}()},
\code{\link{getGenoPred}()},
\code{\link{getHerit}()},
\code{\link{getVar}()},
\code{\link{plot.fitMod}()}
}
\concept{Spatial modeling}
